+++ src/GUI/Categories/LanguageNode.cpp
@@ -9,20 +9,34 @@ bool LanguageNodeUI::init()
     if (!CategoryNode::init())
         return false;
 
-    std::vector<std::string> langs = { "en-AU.json", "ja-JP.json", "de-DE.json", "ru-RU.json", "vi-VI.json", "id-ID.json" };
+    auto langs = LocalisationManager::get()->getAllLanguageFiles();
+    float height = (std::ceil((langs.size() / 2) / 2) * 2) * 2 * (60 + 2.5f);
 
     int x = 0;
     int y = 0;
+    scroll->m_contentLayer->setContentHeight(height);
+
+    std::vector<LanguageNode*> nodes = {};
 
     for (auto lang : langs)
     {
         auto node = LanguageNode::create(lang);
         node->setContentSize(ccp((getContentWidth() - 2.5f * 3) / 2, 60));
-        node->setAnchorPoint(ccp(x, 1));
-        node->setPosition(ccp(x == 0 ? 2.5f : getContentWidth() - 2.5f, getContentHeight() - y * (node->getContentHeight() + 2.5f)) + ccp(0, -25));
         node->updateLayout();
+        nodes.push_back(node);
 
-        this->addChild(node);
+        scroll->m_contentLayer->addChild(node);
+    }
+
+    std::sort(nodes.begin(), nodes.end(), [](LanguageNode* a, LanguageNode* b)
+    {
+        return static_cast<CCFloat*>(a->getUserObject())->getValue() > static_cast<CCFloat*>(b->getUserObject())->getValue();
+    });
+
+    for (auto node : nodes)
+    {
+        node->setAnchorPoint(ccp(x, 1));
+        node->setPosition(ccp(x == 0 ? 2.5f : getContentWidth() - 2.5f, (height) - (y * (node->getContentHeight() + 2.5f))));
 
         x++;
 
@@ -34,4 +48,12 @@ bool LanguageNodeUI::init()
     }
 
     return true;
+}
+
+void LanguageNodeUI::updateUI()
+{
+    scroll->moveToTop();
+    scroll->setTouchEnabled(true);
+    scrollbar->setVisible(false);
+    scrollbar->setDisabled(true);
 }